Split OGL_E1 header-only into .hpp and .cpp
