+++ engines/First OGL Engine/include/SkyboxRenderer.hpp
@@ -1,7 +1,7 @@
 #ifndef OGL_E1_SKYBOXRENDERER_H
 #define OGL_E1_SKYBOXRENDERER_H
 
-#include "First OGL Engine/Utils/EngineTypes.hpp"
+#include "Utils/EngineTypes.hpp"
 #include "Object.hpp"
 #include "Shader.hpp"
 #include "Scene.hpp"
