+++ example/26-ShadowMapping.cpp
@@ -5,7 +5,7 @@
 #include "ShadowMap.hpp"
 #include "CameraCubemap.hpp"
 #include "FrameBufferObject.hpp"
-#include "First OGL Engine/OGL_E1.hpp"
+#include "OGL_E1.hpp"
 
 const int skyboxTextureID = 15;
 const int shadowMapTextureID = 20;
